requirements.txt
absl-py==2.2.2
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
flask==3.0.3
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
gast==0.4.0
google-auth==2.40.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
huggingface-hub==0.31.4
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
markdown==3.7
markupsafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
pandas==2.0.3
pillow==10.4.0
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.2.3
werkzeug==3.0.6
wrapt==1.17.2